requirements.txt
flask>=3.0.0
python-dotenv>=1.0.0
werkzeug>=3.0.0
PyPDF2>=3.0.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
google-generativeai>=0.8.0
